requirements.txt
discord.py>=2.3.0
faster-whisper>=1.0.3
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
